LayoutTests/js/symbol-prototype-is-ordinary-object-expected.txt

This tests that Symbol.prototype object is ordinary object (not Symbol wrapper object).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Symbol.prototype.toString.call(Symbol.prototype) threw exception TypeError: Type error.
PASS Symbol.prototype.toString.call(Symbol.iterator) is "Symbol(Symbol.iterator)"
PASS successfullyParsed is true

TEST COMPLETE
